Bug fix: Scheduler wrongly found some hands to be identical
The old test was rather heuristic and adequate for random distributions
in a batch.  But it was not good enough for distributions that are explicitly
generated to be similar, such as distributions with two given hands
(dummy and a defender) and other hands consistent with the bidding.

The heuristic test is now better, and it is followed by a rigorous test
among the remaining candidates.