Fix Linux memory size detection
The old coding tried to parse the output of `free -k`, but that output
has changed from 4 lines (headers plus 3 lines of actual data) to 3
lines (header plus 2 lines) some years ago. The tail|head|awk pipe now
reads the wrong field, which is now the amount of free swap.

Fix by querying sysconf() for the physical memory, and use half of that.
There is another sysconf() item for the "available" memory, but that
number is not useful, because it excludes memory that is used only for
buffers/caching, which is effectively free to be used.